Create node-server
A localhost server, very handy if you don't want to install XAMPP, and just want a few simple files instead of repositories in your local server. Feel free to fork this for yourself, have a go! Also, where it says "index.html", replace it with your file name, and remember to keep the file relative to this server file. You can change around the ports if you want, that's your choice. Have a great day!